Split out test_headers
